collectors/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
confluent-kafka==1.6.1
cryptography==3.4.6
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.6.0
dnspython==2.1.0
docopt==0.6.2
entrypoints==0.3
greenlet==1.0.0
html5lib==1.1
idna==2.10
importlib-metadata==3.4.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kafka-python==2.0.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
ngxtop==0.0.3
notebook==6.2.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
pydruid==0.6.2
Pygments==2.8.0
PyJWT==1.7.1
pymongo==3.11.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
python-slugify==4.0.1
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
random-object-id==2.0.0
requests==2.25.1
schedule==1.0.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.4.3
SQLAlchemy-Paginator==0.2
sqlparse==0.2.4
tabulate==0.8.7
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
tornado==6.1
traitlets==4.3.3
Tumulus==0.2.1
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
